requirements.txt
pyserial==3.5
pandas==1.3.4
pydantic==1.8.2